format: divided Cards into CardList and CardPopover partials
