+++ src/types/types.ts
@@ -1,4 +1,5 @@
-import type { CLASSES, EVENTS, RARITY_ID } from '@/utils/constants'
+import type CardPopover from '@/components/partials/CardPopover.vue'
+import type { CLASSES, EVENT_TYPES, EVENTS, RARITY_ID } from '@/utils/constants'
 
 export type HearthstoneCard = {
   id: number
@@ -33,7 +34,15 @@ export type RarityId = (typeof RARITY_ID)[RarityIdKey]
 export type EventKey = keyof typeof EVENTS
 export type Event = (typeof EVENTS)[EventKey]
 
+export type EventTypeKey = keyof typeof EVENT_TYPES
+export type EventType = (typeof EVENT_TYPES)[EventTypeKey]
+
 export type QueryParam = {
   key: keyof HearthstoneCard | 'page'
   value: string | number | boolean
+}
+
+export interface CardPopoverRef extends InstanceType<typeof CardPopover> {
+  popoverShow(event: any, image: string): void
+  popoverHide(event: any): void
 }
